requirements.txt
numpy==2.0.2
pydexcom==0.4.1
pyobjc_framework_Cocoa==11.0
pyobjc_framework_Security==11.0
rumps==0.4.0
setuptools==58.0.4
matplotlib==3.9.4
keyring==24.3.1
py2app==0.28